Change from usize to u64
On [random-access-storage](https://github.com/datrs/random-access-storage) issue
https://github.com/datrs/random-access-storage/issues/6 has changed all
types from `usize` into `u64` to be able to handle more than 4gbs on
32bits systems.

> usize is 32 bits on a 32 bit system, so the storage would be limited
to 4GB on such a system.

When changing `random-access-storage` on `hypercore` (tracking:
https://github.com/datrs/hypercore/pull/100) one of the things I've
noticed is that it wold be easier to also make `flat-tree` use `u64` to
integrate with `random-access-storage`. Very likely, this also means
that we would need to bump to use more than 32Gb storages on
`flat-tree`.

I've simply changed all declarations of usize into u64. All tests are
passing.